PO-2087:
1. Changed 'version' type in legacy request and response from Integer to BigInteger
2. Renamed tests with better descriptions

+++ src/main/java/uk/gov/hmcts/opal/dto/legacy/AddPaymentTermsLegacyResponse.java
@@ -13,6 +13,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import uk.gov.hmcts.opal.util.LocalDateAdapter;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Data
@@ -29,7 +30,7 @@ public class AddPaymentTermsLegacyResponse {
 
     @XmlElement(name = "version")
     @NotNull
-    private Integer version;
+    private BigInteger version;
 
     @XmlElement(name = "payment_terms")
     @NotNull
